// src/hotels/interfaces/hotel.interface.ts
/**
 * HotelS interface
 */
export interface Hotel {
  value: number;
  id: number;
  name: string;
  cnpj: string;
  email: string;
  floors: string;
  category: string;
  phone: string;
  note: string;
  address_zip: string;
  address_street: string;
  address_number: string;
  address_district: string;
  address_city: string;
}